Add super-simple authentication
This method is kind of dumb but it protects against
spammy automated / malicious traffic.

We'll do a more "complete" version of authentication
in the near future.

This commit also removes a "trotter-api" executable
that was sitting in the directory. Looks like it
was pretty old since it only has 1 route when you
activate it.